requirements.txt
aiogram==2.22.1
python-dotenv==0.20.0
requests==2.32.3